+++ airflow/post_install.sh
@@ -12,6 +12,17 @@ python $AIRFLOW_HOME/utils/set_default_variables.py \
     temp_bucket="{{ var.value.TEMP_BUCKET }}"  \
     pipeline_bucket="{{ var.value.PIPELINE_BUCKET }}" \
     pipeline_dataset="{{ var.value.PIPELINE_DATASET }}" \
-    ecuador_vms_gcs_path="gs://vms-gfw/ecuador/api" \
+    ecuador_vms_filtered_gcs_path="gs://vms-gfw/ecuador/api" \
+
+python $AIRFLOW_HOME/utils/set_default_variables.py \
+    --force docker_image=$1 \
+    pipe_vms_ecuador \
+    dag_install_path="${THIS_SCRIPT_DIR}" \
+    docker_run="{{ var.value.DOCKER_RUN }}" \
+    project_id="{{ var.value.PROJECT_ID }}" \
+    temp_bucket="{{ var.value.TEMP_BUCKET }}"  \
+    pipeline_bucket="{{ var.value.PIPELINE_BUCKET }}" \
+    pipeline_dataset="{{ var.value.PIPELINE_DATASET }}" \
+    ecuador_vms_filtered_gcs_path="gs://vms-gfw/ecuador/api" \
 
 echo "Installation Complete"