+++ CHANGES.md
@@ -8,9 +8,17 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 
 ## [Unreleased]
 
+## v2.0.0 - 2021-03-31
+
+### Added
+
+* [Data Pipeline/PIPELINE-8/PIPELINE-89](https://globalfishingwatch.atlassian.net/browse/PIPELINE-89): Adds
+  normalization query, schema and script to have the information prepared for
+  the pipeline.
+
 ## v1.0.0 - 2020-08-21
 
 ### Added
 
 * [Data Pipeline/PIPELINE-8/PIPELINE-48](https://globalfishingwatch.atlassian.net/browse/PIPELINE-48): Adds
-  * the fetch script to get Ecuador data daily basis.
+  the fetch script to get Ecuador data daily basis.